First attempt to show footnotes in a popup
In modern browsers, we rely on anchor positioning as a progressive enhancement. In older browsers, it’s an absolutely positioned element.

All footnote formatting, links, etc., as well as the complete list of footnotes at the end of the blog post, are preserved. 

No additional JS needed. All necessary anchors are added on build.

There are some limitations in styling (e.g., it’s not that easy to add a pointer to the popup) because the popup position varies depending on the available space in the browser to display the popup. We can probably use the new container type (`anchored()`) as a progressive enhancement to address that issue.