First attempt to show footnotes in a pop-up (#125)
In modern browsers, we rely on anchor positioning as a progressive enhancement. In older browsers, it’s an absolutely positioned element.

All footnote formatting, links, etc., as well as the complete list of footnotes at the end of the blog post, are preserved. 

No additional JS needed. All necessary anchors are added on build.

There are some limitations in styling (e.g., it’s not that easy to add a pointer to the pop-up) because the pop-up position varies depending on the available space in the browser for displaying the pop-up. We can probably use the new container type (`anchored()`) as a progressive enhancement to address that issue.

* Better styles for older browsers